dapp/ai/computer_vision/requirements.txt
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --find-links=.
#
--find-links https://felipefg.github.io/pip-wheels-riscv/wheels/

    # via requests
numpy==1.24.3
    # via
    #   -r requirements.in
    #   tflite-runtime
pillow==9.5.0
    # via -r requirements.in
tflite-runtime==2.12.0
    # via -r requirements.in